Rename userGroups.jsx to groupCards.jsx
The two exported values are independent of user information, so remove
'user' from the filename.
