+++ static-site/src/components/splash/groupCards.jsx
