+++ static-site/src/pages/groups.jsx
@@ -4,7 +4,7 @@ import { SmallSpacer, HugeSpacer, FlexCenter } from "../layouts/generalComponent
 import * as splashStyles from "../components/splash/styles";
 import { fetchAndParseJSON } from "../util/datasetsHelpers";
 import DatasetSelect from "../components/Datasets/dataset-select";
-import { GroupCards } from "../components/splash/userGroups";
+import { GroupCards } from "../components/splash/groupCards";
 import GenericPage from "../layouts/generic-page";
 import { UserContext } from "../layouts/userDataWrapper";
 import { DataFetchErrorParagraph } from "../components/splash/errorMessages";
